[SystemInfo] Implement storage on mobile
