+++ system_info/system_info.gyp
@@ -75,6 +75,8 @@
         'system_info_sim_mobile.cc',
         'system_info_storage.cc',
         'system_info_storage.h',
+        'system_info_storage_desktop.cc',
+        'system_info_storage_mobile.cc',
         'system_info_utils.cc',
         'system_info_utils.h',
         'system_info_wifi_network.cc',
